fix: orders initial migration
- Deleted the orders/models/ directory because the command 'makemigrations' wasn't detecting changes in those files
- Created the orders/models.py file and created the initial migration
